+++ packages/imperative/src/cmd/src/CommandProcessor.ts
@@ -614,7 +614,7 @@ export class CommandProcessor {
             };
             try {
                 if (handlerParms.arguments.showInputsOnly) {
-                    this.showInputsOnly(response, handlerParms);
+                    this.showInputsOnly(handlerParms);
                 } else {
                     await handler.process(handlerParms);
                 }
@@ -645,60 +645,73 @@ export class CommandProcessor {
 
             let bufferedStdOut = Buffer.from([]);
             let bufferedStdErr = Buffer.from([]);
-            this.log.debug("Attempting to invoke %d chained handlers for command: '%s'", this.definition.chainedHandlers.length,
-                this.definition.name);
-            for (let chainedHandlerIndex = 0; chainedHandlerIndex < this.definition.chainedHandlers.length; chainedHandlerIndex++) {
-                const chainedHandler = this.definition.chainedHandlers[chainedHandlerIndex];
-                this.log.debug("Loading chained handler '%s' (%d of %d)",
-                    chainedHandler.handler, chainedHandlerIndex + 1, this.definition.chainedHandlers.length);
-                const handler: ICommandHandler = this.attemptHandlerLoad(response, chainedHandler.handler);
-                if (handler == null) {
-                    // if the handler load failed
-                    this.log.fatal("failed to load a chained handler! aborting chained handler sequence.");
-                    return this.finishResponse(response);
-                }
-                this.log.debug("Constructing new response object for handler '%s': silent?: %s. json?: %s",
-                    chainedHandler.handler, chainedHandler.silent + "", params.arguments[Constants.JSON_OPTION] + "");
-                chainedResponse = this.constructResponseObject({
-                    arguments: params.arguments,
-                    silent: chainedHandler.silent,
-                    responseFormat: params.arguments[Constants.JSON_OPTION] ? "json" : "default"
+            if (preparedArgs.showInputsOnly) {
+                this.showInputsOnly({
+                    response,
+                    arguments: preparedArgs,
+                    positionals: preparedArgs._,
+                    definition: this.definition,
+                    fullDefinition: this.fullDefinition,
+                    stdin: this.getStdinStream()
                 });
-
-                // make sure the new chained response preserves output
-                chainedResponse.bufferStdout(bufferedStdOut);
-                chainedResponse.bufferStderr(bufferedStdErr);
-                try {
-                    await handler.process({
-                        response: chainedResponse,
-                        arguments: ChainedHandlerService.getArguments(
-                            this.mCommandRootName,
-                            this.definition.chainedHandlers,
-                            chainedHandlerIndex,
-                            chainedResponses,
-                            preparedArgs,
-                            this.log
-                        ),
-                        positionals: preparedArgs._,
-                        definition: this.definition,
-                        fullDefinition: this.fullDefinition,
-                        stdin: this.getStdinStream(),
-                        isChained: true
+                response.succeeded();
+                response.endProgressBar();
+                return this.finishResponse(response);
+            } else {
+                this.log.debug("Attempting to invoke %d chained handlers for command: '%s'", this.definition.chainedHandlers.length,
+                    this.definition.name);
+                for (let chainedHandlerIndex = 0; chainedHandlerIndex < this.definition.chainedHandlers.length; chainedHandlerIndex++) {
+                    const chainedHandler = this.definition.chainedHandlers[chainedHandlerIndex];
+                    this.log.debug("Loading chained handler '%s' (%d of %d)",
+                        chainedHandler.handler, chainedHandlerIndex + 1, this.definition.chainedHandlers.length);
+                    const handler: ICommandHandler = this.attemptHandlerLoad(response, chainedHandler.handler);
+                    if (handler == null) {
+                        // if the handler load failed
+                        this.log.fatal("failed to load a chained handler! aborting chained handler sequence.");
+                        return this.finishResponse(response);
+                    }
+                    this.log.debug("Constructing new response object for handler '%s': silent?: %s. json?: %s",
+                        chainedHandler.handler, chainedHandler.silent + "", params.arguments[Constants.JSON_OPTION] + "");
+                    chainedResponse = this.constructResponseObject({
+                        arguments: params.arguments,
+                        silent: chainedHandler.silent,
+                        responseFormat: params.arguments[Constants.JSON_OPTION] ? "json" : "default"
                     });
-                    const builtResponse = chainedResponse.buildJsonResponse();
-                    chainedResponses.push(builtResponse.data);
-                    // save the stdout and stderr to pass to the next chained handler (if any)
-                    bufferedStdOut = builtResponse.stdout;
-                    bufferedStdErr = builtResponse.stderr;
 
-                } catch (processErr) {
-                    this.handleHandlerError(processErr, chainedResponse, chainedHandler.handler);
+                    // make sure the new chained response preserves output
+                    chainedResponse.bufferStdout(bufferedStdOut);
+                    chainedResponse.bufferStderr(bufferedStdErr);
 
-                    // Return the failed response to the caller
-                    return this.finishResponse(chainedResponse);
+                    try {
+                        await handler.process({
+                            response: chainedResponse,
+                            arguments: ChainedHandlerService.getArguments(
+                                this.mCommandRootName,
+                                this.definition.chainedHandlers,
+                                chainedHandlerIndex,
+                                chainedResponses,
+                                preparedArgs,
+                                this.log
+                            ),
+                            positionals: preparedArgs._,
+                            definition: this.definition,
+                            fullDefinition: this.fullDefinition,
+                            stdin: this.getStdinStream(),
+                            isChained: true
+                        });
+                        const builtResponse = chainedResponse.buildJsonResponse();
+                        chainedResponses.push(builtResponse.data);
+                        // save the stdout and stderr to pass to the next chained handler (if any)
+                        bufferedStdOut = builtResponse.stdout;
+                        bufferedStdErr = builtResponse.stderr;
+                    } catch (processErr) {
+                        this.handleHandlerError(processErr, chainedResponse, chainedHandler.handler);
+
+                        // Return the failed response to the caller
+                        return this.finishResponse(chainedResponse);
+                    }
                 }
             }
-
             this.log.info(`Chained handlers for command "${this.definition.name}" succeeded.`);
             response.succeeded();
             response.endProgressBar();
@@ -716,7 +729,7 @@ export class CommandProcessor {
      * @returns
      * @memberof CommandProcessor
      */
-    private showInputsOnly(response: IHandlerResponseApi, commandParameters: IHandlerParameters) {
+    private showInputsOnly(commandParameters: IHandlerParameters) {
 
         /**
          * Determine if we should display secure values.  If the ENV variable is set to true,
@@ -799,8 +812,8 @@ export class CommandProcessor {
          * Show warning if we censored output and we were not instructed to show secure values
          */
         if (censored && !showSecure) {
-            response.console.errorHeader("Some inputs are not displayed");
-            response.console.error(
+            commandParameters.response.console.errorHeader("Some inputs are not displayed");
+            commandParameters.response.console.error(
                 `Inputs below may be displayed as '${ConfigConstants.SECURE_VALUE}'. ` +
                 `Properties identified as secure fields are not displayed by default.\n\n` +
                 `Set the environment variable ` +