+++ packages/cli/CHANGELOG.md
@@ -1,7 +1,11 @@
 # Change Log
 All notable changes to the Zowe CLI package will be documented in this file.
 
-## Recent Changes
+## `8.15.1`
+
+- BugFix: Fixed the `--show-inputs-only` option on commands with chained command handlers, such as `zowe zos-files copy data-set-cross-lpar`. [#2446](https://github.com/zowe/zowe-cli/issues/2446)
+
+## `8.15.0`
 
 - Enhancement: Added the `overwrite` parameter to the `Copy.DataSet()` command to allow for overwriting all members of a target data set with source data set members. [#2450] (https://github.com/zowe/zowe-cli/pull/2450)