+++ packages/imperative/CHANGELOG.md
@@ -5,8 +5,14 @@ All notable changes to the Imperative package will be documented in this file.
 ## Recent Changes
 
 - BugFix: When in daemon mode, the user would not see Imperative initialization errors, but now the errors are passed back to the user's terminal window. [#1875] (https://github.com/zowe/zowe-cli/issues/1875).
+- Enhancement: Added a favicon to the Web Help that displays in browser tabs. [#801] (https://github.com/zowe/zowe-cli/issues/801)
+
+## `8.15.1`
+
+- BugFix: Fixed the `--show-inputs-only` option on commands with chained command handlers. [#2446](https://github.com/zowe/zowe-cli/issues/2446)
 
 ## `8.14.1`
+
 - BugFix: Fixed help text example section's wrapping issue where the first line of the description is wrapped differently than the rest of the lines. [#1945] (https://github.com/zowe/zowe-cli/issues/1945).
 
 ## `8.14.0`
@@ -44,7 +50,6 @@ All notable changes to the Imperative package will be documented in this file.
 ## `8.10.0`
 
 - BugFix: Modified location of Proxy-Authorization header to be located in the agent instead of the request. [#2389](https://github.com/zowe/zowe-cli/issues/2389)
-
 
 ## `8.8.3`